Implement hardware agnostic training by removing .to usages
https://pytorch-lightning.readthedocs.io/en/latest/accelerators/accelerator_prepare.html

+++ ldm/modules/diffusionmodules/model.py
@@ -22,8 +22,10 @@ def get_timestep_embedding(timesteps, embedding_dim):
     half_dim = embedding_dim // 2
     emb = math.log(10000) / (half_dim - 1)
     emb = torch.exp(torch.arange(half_dim, dtype=torch.float32) * -emb)
+    # FIXME: Hardware agnostic training
+    # should this be .type_as(timesteps) ?
     emb = emb.to(device=timesteps.device)
-    emb = timesteps.float()[:, None] * emb[None, :]
+    emb = timesteps[:, None] * emb[None, :]
     emb = torch.cat([torch.sin(emb), torch.cos(emb)], dim=1)
     if embedding_dim % 2 == 1:  # zero pad
         emb = torch.nn.functional.pad(emb, (0, 1, 0, 0))
@@ -488,6 +490,8 @@ class Encoder(nn.Module):
         temb = None
 
         # downsampling
+        # print("DBG model.py forward x type ", x.type())
+        # Fails here due to wrong weight type
         hs = [self.conv_in(x)]
         for i_level in range(self.num_resolutions):
             for i_block in range(self.num_res_blocks):
